+++ functions/src/notification_validation.ts
@@ -1,0 +1,53 @@
+const NOTIFICATION_TYPE_ID_PATTERN = /^[A-Za-z0-9_-]{1,64}$/;
+
+export type NotificationGender = "male" | "female" | "other";
+
+type DynamicNotificationType = {
+  messageType: "dynamic";
+  quotesCollections: Record<string, unknown>;
+};
+
+type StaticNotificationType = {
+  messageType: "static";
+  staticTitle: string;
+  staticBody: string;
+};
+
+type ValidNotificationType = DynamicNotificationType | StaticNotificationType;
+
+export function isValidNotificationTypeId(value: unknown): value is string {
+  return (
+    typeof value === "string" && NOTIFICATION_TYPE_ID_PATTERN.test(value)
+  );
+}
+
+export function normalizeNotificationGender(
+  value: unknown,
+): NotificationGender {
+  return value === "male" || value === "female" || value === "other"
+    ? value
+    : "other";
+}
+
+export function hasValidNotificationTypeSchema(
+  value: unknown,
+): value is ValidNotificationType {
+  if (value === null || typeof value !== "object" || Array.isArray(value)) {
+    return false;
+  }
+  const data = value as Record<string, unknown>;
+  if (data.messageType === "dynamic") {
+    return (
+      data.quotesCollections !== null &&
+      typeof data.quotesCollections === "object" &&
+      !Array.isArray(data.quotesCollections)
+    );
+  }
+  if (data.messageType === "static") {
+    return (
+      typeof data.staticTitle === "string" &&
+      typeof data.staticBody === "string"
+    );
+  }
+  return false;
+}
